Add freqamp-PM syntax. Refactor & optimize oscillator.
Language changes:
* Add "p,f[...]" frequency-amplified PM modulator list syntax.
  Unlike for normal PM, the modulation index (amplitudes) are
  multiplied by the carrier frequency, scaled down so ~632 Hz
  gives normal amplitude.

Refactor oscillator, split "phasor" type with new
buffer-filling function, make the main run functions
use passed pre-filled integer phase increase and offset
buffers, instead of buffers for frequency and phase offset
in floating-point. The splitting of the processing loop turns
out to increase performance signficantly, in additon.

Add ~632 Hz constant to math.h, used as a sensible
middle-point in the frequency range (the geometric
mean of 20 and 20000, common for audio equipment.)

Change log:
 * 2022-02-20: Implement "p+f[]" syntax, refactor.
 * 2022-02-21: Update documentation.
 * 2022-02-22: Tweak scripts & use the new syntax:
       'examples/errorsignal.sgs'
       'examples/tests/vibrato-pm.sgs'  (new file)
       'devtests/music04.sgs'
 * 2022-02-23: Change naming in the documentation.
       (frequency-multiplied -> frequency-linked.)
       Adjust -p printout, use 3-letter mod abbrs.
 * 2022-06-14: Rename 'phasor' to 'freqor'  (as it
       provides non-integrated values). Remove env
       version of oscillator code, split out mixer
       functions for buffer handling including amp
       and use in generator.c - same output/result
       and also tests faster on my x86-64 laptops.
 * 2022-06-15: Simplify new generator.c code more.
 * 2022-07-11: Merge new syntax changes, fPM usage
       adjusted to "p,f[]" matching later version.
 * 2022-08-24: Simplify the oscillator more making
       a plain phasor out of the freqor. Removes a
       buffer from the generator code, works well.
       Set oscillator phase as uint32_t in parser.
 * 2022-12-20: Fix 'p' bug (there
       since 2018) on 32-bit x86.

+++ script.h
@@ -49,7 +49,7 @@ typedef struct SGS_ScriptOpData {
 	SGS_Ramp *freq, *freq2;
 	SGS_Ramp *amp, *amp2;
 	SGS_Ramp *pan;
-	float phase;
+	uint32_t phase;
 	/* node adjacents in operator linkage graph */
 	SGS_ScriptListData *mods;
 	/* for conversion */
